requirements.txt
discord.py==2.1.0
requests==2.25.1
youtube-search-python==1.6.6
googlesearch-python==1.1.0
youtube-search==2.1.2


















